# backend/src/swagger.yaml
openapi: "3.0.3"
info:
  title: Plotra API
  description: 소설 창작 AI 보조 플랫폼 - REST API
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /api/projects/{projectId}/world-items:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [World Items]
      summary: List world items
      operationId: listWorldItems
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorldItem" }
    post:
      tags: [World Items]
      summary: Create world item
      operationId: createWorldItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name: { type: string }
                category:
                  type: string
                  enum: [region, nation, city, race, organization, magic_system, technology, rule, religion, language, other]
                content: { type: string }
                summary: { type: string }
                tags:
                  type: array
                  items: { type: string }
                metadata: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldItem" }
  /api/projects/{projectId}/world-items/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [World Items]
      summary: Get world item
      operationId: getWorldItem
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldItem" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [World Items]
      summary: Update world item
      operationId: updateWorldItem
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                category:
                  type: string
                  enum: [region, nation, city, race, organization, magic_system, technology, rule, religion, language, other]
                content: { type: string }
                summary: { type: string }
                tags:
                  type: array
                  items: { type: string }
                metadata: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldItem" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [World Items]
      summary: Delete world item (soft)
      operationId: deleteWorldItem
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── World Item Links ───
  /api/projects/{projectId}/world-items/{id}/links:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [World Item Links]
      summary: List world item links
      operationId: listWorldItemLinks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorldItemLink" }
    post:
      tags: [World Item Links]
      summary: Create world item link
      operationId: createWorldItemLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetId]
              properties:
                targetId: { type: string, format: uuid }
                description: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorldItemLink" }
  /api/projects/{projectId}/world-items/{id}/links/{linkId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/LinkId"
    delete:
      tags: [World Item Links]
      summary: Delete world item link
      operationId: deleteWorldItemLink
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── World Item ↔ Characters ───
  /api/projects/{projectId}/world-items/{id}/characters:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
    get:
      tags: [World Item Characters]
      summary: List character links for world item
      operationId: listWorldItemCharacters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CharacterWorldItem" }
    post:
      tags: [World Item Characters]
      summary: Link character to world item
      operationId: createWorldItemCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [characterId]
              properties:
                characterId: { type: string, format: uuid }
                description: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CharacterWorldItem" }
  /api/projects/{projectId}/world-items/{id}/characters/{linkId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/LinkId"
    delete:
      tags: [World Item Characters]
      summary: Unlink character from world item
      operationId: deleteWorldItemCharacter
      responses:
        "200": { $ref: "#/components/responses/Deleted" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Plots ───
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Encrypted JWT token"
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LinkId:
      name: linkId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    Deleted:
      description: Deleted
      content:
        application/json:
          schema:
            type: object
            properties:
              deleted: { type: boolean }
  schemas:
...
    WorldItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        name: { type: string }
        category:
          type: string
          enum: [region, nation, city, race, organization, magic_system, technology, rule, religion, language, other]
        content: { type: string }
        summary: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    WorldItemLink:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        sourceId: { type: string, format: uuid }
        targetId: { type: string, format: uuid }
        description: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CharacterWorldItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        characterId: { type: string, format: uuid }
        worldItemId: { type: string, format: uuid }
        description: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...